requirements.txt
aiohttp==3.10.5
alembic==1.13.2
atomicwrites==1.4.1
fastapi[standard]==0.114.2
psutil==6.0.0
pydantic-settings==2.4.0
pydantic==2.9.1
pyjwt==2.9.0
pyotp==2.9.0
python-multipart==0.0.9
pytz==2024.1
qrcode==7.4.2
semver==3.0.2
sqlalchemy==2.0.34
sqlmodel==0.0.22
suntime==1.3.2
uvicorn==0.30.6
